requirements.txt
flask
openai
openai-agents
python-dotenv
